Agent: Move master/threading_utils.py -> utils/threading.py
Both create_daemon_thread() and interruptable_iter() will need to be
used outside of the master.
