+++ monkey/infection_monkey/master/exploiter.py
@@ -7,8 +7,7 @@ from typing import Callable, Dict, List
 
 from infection_monkey.i_puppet import ExploiterResultData, IPuppet
 from infection_monkey.model import VictimHost
-
-from .threading_utils import run_worker_threads
+from infection_monkey.utils.threading import run_worker_threads
 
 QUEUE_TIMEOUT = 2
